add test for this thing
